fix custom wrapper
