+++ packages/cadl-typescript/src/modelUtils.ts
@@ -43,11 +43,13 @@ import {
   Schema,
   SchemaContext
 } from "@azure-tools/rlc-common";
+import { getResourceOperation } from "@cadl-lang/rest";
 import {
   getHeaderFieldName,
   getPathParamName,
   getQueryParamName,
-  isStatusCode
+  isStatusCode,
+  HttpOperation
 } from "@cadl-lang/rest/http";
 import { getPagedResult } from "@azure-tools/cadl-azure-core";
 
@@ -904,4 +906,42 @@ export function getFormattedPropertyDoc(
     return `${propertyDoc}${sperator}${enhancedDocFromType}`;
   }
   return propertyDoc ?? enhancedDocFromType;
+}
+
+export function getBodyType(program: Program, route: HttpOperation) {
+  let bodyModel = route.parameters.bodyType;
+  if (bodyModel && bodyModel.kind === "Model" && route.operation) {
+    const resourceType = getResourceOperation(
+      program,
+      route.operation
+    )?.resourceType;
+    if (resourceType && route.responses && route.responses.length > 0) {
+      const resp = route.responses[0];
+      if (resp && resp.responses && resp.responses.length > 0) {
+        const responseBody = resp.responses[0]?.body;
+        if (responseBody) {
+          const bodyTypeInResponse = getEffectiveModelFromType(
+            program,
+            responseBody.type
+          );
+          // response body type is reosurce type, and request body type (if templated) contains resource type
+          if (
+            bodyTypeInResponse === resourceType &&
+            bodyModel.templateArguments &&
+            bodyModel.templateArguments.some((it) => {
+              return it.kind === "Model" || it.kind === "Union"
+                ? it === bodyTypeInResponse
+                : false;
+            })
+          ) {
+            bodyModel = resourceType;
+          }
+        }
+      }
+    }
+    if (resourceType && bodyModel.name === "") {
+      bodyModel = resourceType;
+    }
+  }
+  return bodyModel;
 }
